refactor(folder) Fix small typo on components folder
