+++ src/index.ts
@@ -1,5 +1,5 @@
 export * from './calendar.model';
 export { CalendarModule } from './calendar.module';
 export { CalendarController } from './calendar.controller';
-export { CalendarWeekTitle } from './compopnents/week-title'
+export { CalendarWeekTitle } from './components/week-title'
 export * from './calendar';
